+++ playerinfo.php
@@ -5,21 +5,17 @@ include_once "templates/header.php" ;
 require_once("config/db.php");
 
 $connection = mysql_connect($db_host, $db_username, $db_password);
-		if(!$connection)
-        {
-            die ("Could not connect to the database: <br />". mysql_error());
-        }
-        
-        
-        $db_select = mysql_select_db($db_database, $connection);
-        if (!$db_select)
-        {
-            die ("Could not select the database: <br />". mysql_error());
-        }
+if(!$connection){
+    die ("Could not connect to the database: <br />". mysql_error());
+}
+
+$db_select = mysql_select_db($db_database, $connection);
+if (!$db_select){
+    die ("Could not select the database: <br />". mysql_error());
+}
 		
 		
-if (isset($_GET["playerid"]))
-{
+if (isset($_GET["playerid"])){
 	$player = $_GET["playerid"];
 }