+++ templates/header.php
@@ -12,13 +12,13 @@
     <meta name="viewport" content="width=device-width">
     <link rel="stylesheet" href="/assets/css/normalize.css">
     <link rel="stylesheet" href="/assets/css/tnt.css" />
-    <script src="assets/js/vendor/modernizr-2.6.2.min.js"></script>
+    <script src="/assets/js/vendor/modernizr-2.6.2.min.js"></script>
   </head>
 
   <body>
     <table width="100%" border="1" cellpadding="10">
       <tr>
         <td align="center" valign="top" width="15%" height="50px">
-          <a href="index.php"><img src="/assets/img/tntbox.JPG"></a>
+          <a href="/index.php"><img src="/assets/img/tntbox.JPG"></a>
         </td>
         <td valign="top" width="85%" rowspan="2">